Moved db call back to context provider to avoid clientside rendering bug
Relates #42
