+++ app/(dashboard)/conversations/page.tsx
@@ -4,7 +4,6 @@ import PlusIcon from '@/components/icons/PlusIcon';
 import ConversationProvider from '@/components/messaging/ConversationProvider';
 import ConversationsList from '@/components/messaging/ConversationsList';
 import CurrentConversation from '@/components/messaging/CurrentConversation';
-import selectUserConversationsandItemNames from '@/utils/messaging/selectUserConversationsandItemNames';
 import { createServerComponentClient } from '@supabase/auth-helpers-nextjs';
 import { cookies } from 'next/headers';
 
@@ -13,14 +12,9 @@ const Conversations = async () => {
   const { data } = await supabase.auth.getSession();
   const userId = data.session?.user.id;
 
-  const conversations = await selectUserConversationsandItemNames(
-    supabase,
-    userId
-  );
-
   return (
     userId && (
-      <ConversationProvider conversations={conversations} userId={userId}>
+      <ConversationProvider userId={userId}>
         <div className='mt-4 flex justify-between px-3 '>
           <button>
             <PlusIcon width={45} height={45} />
