+++ components/messaging/ConversationProvider.tsx
@@ -1,21 +1,21 @@
 'use client';
-import React, { ReactNode, useState } from 'react';
+import React, { ReactNode, useEffect, useState } from 'react';
 import {
   AllConversationsType,
   ConversationCardType,
 } from '@/utils/messaging/messagingTypes';
+import selectUserConversationsandItemNames from '@/utils/messaging/selectUserConversationsandItemNames';
 import useConversation from '../../app/(dashboard)/conversations/useConversation';
 
 const ConversationProvider = ({
   children,
-  conversations,
+  userId,
 }: {
   children: ReactNode;
   userId: string;
-  conversations: AllConversationsType;
 }) => {
   const [allConversations, setAllConversations] =
-    useState<AllConversationsType>(conversations);
+    useState<AllConversationsType>([]);
 
   const [currentConversation, setCurrentConversation] =
     useState<ConversationCardType>({
@@ -29,6 +29,16 @@ const ConversationProvider = ({
         item_name: 'default',
       },
     });
+
+  useEffect(() => {
+    const fetchConversations = async () => {
+      const fetchedConversations =
+        await selectUserConversationsandItemNames(userId);
+
+      setAllConversations(fetchedConversations);
+    };
+    fetchConversations();
+  }, []);
 
   return (
     <useConversation.Provider
